requirements.txt
PyQt5
reportlab
nltk
pytrends
newsapi-python
pillow